requirements.txt
flask
flask_sqlalchemy
flask_login
requests
python-dotenv
pandas
shapely
openpyxl
